requirements.txt
Flask==2.0.1
flask_smorest==0.34.0
PyYAML==5.4.1
gunicorn
gevent==23.9.1
psycopg2-binary==2.9.10
